Updated to latest adapter and runtime 20.91.62.4 and iab code for validation
Added some IAB usage to show Channel API and IAB working with the latest adapter and version.

Right click in the top of the grid (a bit fiddly) to get the inspect option. Once loaded you should see messages appear in the console when you select an option from the wpf list.

When you click on something on the grid line 84 of MainWindow.xaml.cs will hit (you can set a breakpoint to validate).